Even more 64-bit safe values
git-svn-id: svn://svn.code.sf.net/p/playerpro/code/branches/PlayerPRO6@177 f4605116-7794-4d0b-9e48-65ca9bfeedfd

+++ MADDriver.source/MADDriver.Source/Import-Export/669.c
@@ -48,7 +48,7 @@ static inline void mystrcpy( Ptr a, BytePtr b)
 	memmove( a, b + 1, b[ 0]);
 }
 
-static OSErr Convert6692Mad( Ptr	AlienFile, long MODSize, MADMusic	*theMAD, MADDriverSettings *init)
+static OSErr Convert6692Mad( Ptr AlienFile, size_t MODSize, MADMusic *theMAD, MADDriverSettings *init)
 {
 	SixSixNine			*the669;
 	short				i, PatMax, x, z;
